+++ packages/field/src/components/IndexColumn/index.tsx
@@ -1,6 +1,6 @@
 import React, { useContext } from 'react';
 import classnames from 'classnames';
-import { ConfigContext } from 'antd/lib/config-provider/context';
+import { ConfigProvider } from 'antd';
 import './index.less';
 
 /**
@@ -11,7 +11,8 @@ const IndexColumn: React.ForwardRefRenderFunction<any, { border?: boolean; child
   { border = false, children },
   ref,
 ) => {
-  const { getPrefixCls } = useContext(ConfigContext);
+  const { getPrefixCls } = useContext(ConfigProvider.ConfigContext);
+
   const className = getPrefixCls('pro-field-index-column');
   return (
     <div