+++ packages/table/src/component/Alert/index.tsx
@@ -1,23 +1,23 @@
 import React, { useContext } from 'react';
-import { ConfigContext } from 'antd/lib/config-provider/context';
-import { Alert, Space } from 'antd';
+import { Alert, Space, ConfigProvider } from 'antd';
 import './index.less';
 import { useIntl, IntlType } from '@ant-design/pro-provider';
+
+export type AlertRenderType<T> =
+  | ((props: {
+      intl: IntlType;
+      selectedRowKeys: (number | string)[];
+      selectedRows: T[];
+      onCleanSelected: () => void;
+    }) => React.ReactNode)
+  | false;
 
 export interface TableAlertProps<T> {
   selectedRowKeys: (number | string)[];
   selectedRows: T[];
-  alertInfoRender?:
-    | ((props: {
-        intl: IntlType;
-        selectedRowKeys: (number | string)[];
-        selectedRows: T[];
-      }) => React.ReactNode)
-    | false;
+  alertInfoRender?: AlertRenderType<T>;
   onCleanSelected: () => void;
-  alertOptionRender?:
-    | false
-    | ((props: { intl: IntlType; onCleanSelected: () => void }) => React.ReactNode);
+  alertOptionRender?: AlertRenderType<T>;
 }
 
 const defaultAlertOptionRender = (props: { intl: IntlType; onCleanSelected: () => void }) => {
@@ -36,7 +36,7 @@ const TableAlert = <T, U = {}>({
   alertInfoRender = ({ intl }) => (
     <Space>
       {intl.getMessage('alert.selected', '已选择')}
-      <a style={{ fontWeight: 600 }}>{selectedRowKeys.length}</a>
+      {selectedRowKeys.length}
       {intl.getMessage('alert.item', '项')}&nbsp;&nbsp;
     </Space>
   ),
@@ -48,14 +48,16 @@ const TableAlert = <T, U = {}>({
     alertOptionRender &&
     alertOptionRender({
       onCleanSelected,
+      selectedRowKeys,
+      selectedRows,
       intl,
     });
-  const { getPrefixCls } = useContext(ConfigContext);
+  const { getPrefixCls } = useContext(ConfigProvider.ConfigContext);
   const className = getPrefixCls('pro-table-alert');
   if (alertInfoRender === false) {
     return null;
   }
-  const dom = alertInfoRender({ intl, selectedRowKeys, selectedRows });
+  const dom = alertInfoRender({ intl, selectedRowKeys, selectedRows, onCleanSelected });
   if (dom === false || selectedRowKeys.length < 1) {
     return null;
   }
@@ -69,7 +71,6 @@ const TableAlert = <T, U = {}>({
           </div>
         }
         type="info"
-        showIcon
       />
     </div>
   );