+++ packages/table/src/component/ColumnSetting/index.tsx
@@ -1,16 +1,15 @@
 import React, { useContext, useEffect, useRef } from 'react';
 import { useIntl } from '@ant-design/pro-provider';
-import { ConfigContext } from 'antd/lib/config-provider';
 import {
   SettingOutlined,
   VerticalAlignMiddleOutlined,
   VerticalAlignTopOutlined,
   VerticalAlignBottomOutlined,
 } from '@ant-design/icons';
-import { Checkbox, Popover, Tooltip } from 'antd';
+import { Checkbox, Popover, ConfigProvider, Tooltip } from 'antd';
 import { DndProvider } from 'react-dnd';
 import classNames from 'classnames';
-import Backend from 'react-dnd-html5-backend';
+import { HTML5Backend } from 'react-dnd-html5-backend';
 
 import Container, { ColumnsState } from '../../container';
 import { ProColumns } from '../../Table';
@@ -136,7 +135,7 @@ const CheckboxList: React.FC<{
     const newMap = { ...columnsMap };
     const newColumns = [...sortKeyColumns];
     const findIndex = newColumns.findIndex((columnKey) => columnKey === id);
-    if (findIndex === undefined) {
+    if (findIndex < 0) {
       return;
     }
     const index = newColumns[findIndex];
@@ -176,7 +175,7 @@ const CheckboxList: React.FC<{
     );
   });
   return (
-    <DndProvider backend={Backend}>
+    <DndProvider backend={HTML5Backend}>
       {showTitle && <span className={`${className}-list-title`}>{listTitle}</span>}
       {listDom}
     </DndProvider>
@@ -241,10 +240,12 @@ const ColumnSetting = <T, U = {}>(props: ColumnSettingProps<T>) => {
   const localColumns: Omit<ProColumns<any> & { index?: number }, 'ellipsis'>[] =
     props.columns || counter.columns || [];
 
-  const { columnsMap, setColumnsMap, setSortKeyColumns } = counter;
+  const { columnsMap, setColumnsMap } = counter;
 
   useEffect(() => {
-    columnRef.current = JSON.parse(JSON.stringify(columnsMap));
+    if (columnsMap) {
+      columnRef.current = JSON.parse(JSON.stringify(columnsMap));
+    }
   }, []);
 
   /**
@@ -272,7 +273,7 @@ const ColumnSetting = <T, U = {}>(props: ColumnSettingProps<T>) => {
   const indeterminate = selectedKeys.length > 0 && selectedKeys.length !== localColumns.length;
 
   const intl = useIntl();
-  const { getPrefixCls } = useContext(ConfigContext);
+  const { getPrefixCls } = useContext(ConfigProvider.ConfigContext);
   const className = getPrefixCls('pro-table-column-setting');
 
   return (
@@ -296,7 +297,6 @@ const ColumnSetting = <T, U = {}>(props: ColumnSettingProps<T>) => {
           <a
             onClick={() => {
               setColumnsMap(columnRef.current);
-              setSortKeyColumns([]);
             }}
           >
             {intl.getMessage('tableToolBar.reset', '重置')}