+++ tests/card/index.test.tsx
@@ -42,4 +42,26 @@ describe('Field', () => {
     await waitForComponentToPaint(wrapper);
     expect(wrapper.find('.ant-pro-card-body-collapse').exists()).toBeFalsy();
   });
+
+  it('🥩 tabs onChange', async () => {
+    const fn = jest.fn();
+    const wrapper = mount(
+      <ProCard
+        tabs={{
+          onChange: fn,
+        }}
+      >
+        <ProCard.TabPane key="tab1" tab="产品一">
+          内容一
+        </ProCard.TabPane>
+        <ProCard.TabPane key="tab2" tab="产品二">
+          内容二
+        </ProCard.TabPane>
+      </ProCard>,
+    );
+
+    wrapper.find('.ant-pro-card-tabs .ant-tabs-tab').at(1).simulate('click');
+
+    expect(fn).toHaveBeenCalledWith('tab2');
+  });
 });